maybe not this time for code

+++ service/src/ekskul/ekskul.service.ts
@@ -1,9 +1,13 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
+import { UserService } from 'src/user/user.service';
 
 @Injectable()
 export class EkskulService {
-    constructor(private readonly databaseService: DatabaseService) {}
+    constructor(
+        private readonly databaseService: DatabaseService,
+        private readonly userService: UserService
+    ) {}
 
     // Get All Ekskul
     async getAllEkskul(){
@@ -17,21 +21,42 @@ export class EkskulService {
 
     // Get One Ekskul
     async getOneEkskul(id: number){
-        const get = await this.databaseService.connection("ekskul").select("*").where({id: id}).first()
+        const getEkskul = await this.databaseService.connection("ekskul")
+        .select("*")
+        .where({id: id})
+        .first()
+        
+        const getListMember = await this.databaseService.connection("member_ekskul")
+        .innerJoin("users", "users.id", "member_ekskul.user_id")
+        .select({
+            name: "users.name",
+            nis: "users.nis",
+            role: "member_ekskul.role"
+        })
+        .where({ekskul_id: id})
+    
 
-        if(!get) throw new NotFoundException("Ekskul Tidak Ada")
+        if(!getEkskul) throw new NotFoundException("Ekskul Tidak Ada")
 
         return {
             message: "Berhasil Mendapatkan Detail Ekskul",
-            data: get
+            data: {
+                ekskul: getEkskul,
+                member: getListMember
+            }
         }
     }
 
     // Create Ekskul
-    async createEkskul(name: string, logo: Express.Multer.File){
+    async createEkskul(req: { id: number }, name: string, logo: Express.Multer.File){
+        // Apakah Osis
+        const checkRole = await this.userService.getRole(req.id)
+        if(checkRole != "Osis") throw new ForbiddenException("Anda Tidak Berhak")
+
         if(!name || !logo) throw new BadRequestException("Lengkapi Data yang dibutuhkan")
+
         const path = `/uploads/ekskul/logo/${logo.filename}`
-        const create = await this.databaseService.connection("ekskul").insert({ name: name, logo: path}).returning("*")
+        const [create] = await this.databaseService.connection("ekskul").insert({ name: name, logo: path}).returning("*")
 
         return {
             message: "Berhasil Membuat Ekskul",
@@ -40,7 +65,11 @@ export class EkskulService {
     }
 
     // Delete Ekskul
-    async deleteEKskul(id: number){
+    async deleteEkskul(req: { id: number }, id: number){
+        // Apakah Osis
+        const checkRole = await this.userService.getRole(req.id)
+        if(checkRole != "Osis") throw new ForbiddenException("Anda Tidak Berhak")
+
         const del = await this.databaseService.connection("ekskul").delete().where({id: id})
 
         return {
@@ -50,8 +79,15 @@ export class EkskulService {
 
     // Join Ekskul
     async joinEkskul(req: { id: number}, ekskul_id: number){
+        // Apakah Sudah Join
+        const isAlreadyJoin = await this.userService.getRole(req.id, ekskul_id)
+        if(isAlreadyJoin) throw new ConflictException("Anda Sudah Menjadi Anggota")
+
+        // Join
         const join = await this.databaseService.connection("member_ekskul")
-        .insert({ user_id: req.id, ekskul_id: ekskul_id, role: 'member'})
+        .insert({ user_id: req.id, ekskul_id: ekskul_id, role: 'Member'})
+
+
         return {
             message: "Berhasil Join Ekskul"
         }
